2022 day 16: refactor
Move from multiple arrays of single data points to an array of a struct
with these and add helpful methods.

Also achieve a significant speed-up by using the total rate and the number
of open valves as an approximation of the actual opened valves which
makes cloning and sorting them redundant.

The whole thing now needs 150MB and 5s for me.