2022 day 19 part 1 (still wip)
This was a cool idea, but it falls short in multi-leveled predictions
because it only considers one level of prioritization
(which is enough for the first test case
but unfortunately not for the second).
Well, it was fun-ish while it lasted,
but I don't have enough time for this anymore.