2023: enable warnings
I tried to set this for the language server, but only found a way via
`hie.yaml`, which fails spectacularly for my setup because I'm not
using stack or cabal and don't want to learn how to either.

Setting the option just in the `justfile` would be an option, but then I
don't get hlint (via HLS) to highlight it to me and offer a quick
action to remove the unused import (along with all other hints).